Merge of 'mongodb-version' into master
Porting all the mongodb-version changes into master now. Unable to create a formal pull request due to problems with each branches commit history. This merge will now allow for bot information to stay saved for longer instead of just for a day. It also makes the bot use a database to store and query information instead of using a simple file.